Initial Commit of FizzBuzz Kata done in Ruby
Includes Cucumber test of the project using a scenario outline.
FizzBuzz, for those unfamiliar, takes a given number and prints out the
number with special rules. If the number is divisible by 3, it prints
fizz, if it’s divisible by 5 it prints buzz and if it’s divisible by
both 5 and 3, it prints fizzbuzz, else it prints out the number